% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addExpLp.R
\name{addExpLp}
\alias{addExpLp}
\title{adds the expLp to a prediction}
\usage{
addExpLp(prediction, baselineSurvival)
}
\arguments{
\item{prediction}{form a predictPlp}

\item{baselineSurvival}{}
}
\value{
returns the prediction with an extra column
}
\description{
adds the expLp to a prediction
}
