% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustBaselineSurvival.R
\name{AdjustBaselineSurvival}
\alias{AdjustBaselineSurvival}
\title{adjust an Baseline Survival function such that the overal event rate is fixed}
\usage{
AdjustBaselineSurvival(
  plpModel,
  TrainingSet,
  plpData,
  populationSettings,
  timeTofixat,
  proptofixwith,
  intervalSolution = c(-100, 100)
)
}
\arguments{
\item{plpModel}{a plpModel u want to adjust}

\item{TrainingSet}{the training used for fitting the model, is used for adjusting the model}

\item{plpData}{the plpData}

\item{populationSettings}{the populationsettings.}

\item{timeTofixat}{a time in days to fix the eventrate at. has to be in the times of the baseline survival}

\item{proptofixwith}{probabilities to fix}

\item{intervalSolution}{as this function solves an equation is needs an inteval for finding the solution.}
}
\value{
returns a modified Baselinesurvivalfunction
}
\description{
adjust an Baseline Survival function such that the overal event rate is fixed
}
