% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlotfreq.R
\name{boxPlotfreq}
\alias{boxPlotfreq}
\title{Makes a boxplot for the emperical frequency of the outcome.}
\usage{
boxPlotfreq(data, parameter, NoOfCov, Indexes)
}
\arguments{
\item{data}{original dataset}

\item{parameter}{parameters for the model used to generate the datasets}

\item{NoOfCov}{number/place of the covariate u want study with the plot}

\item{Indexes}{matrix with for echt colum a simulation and each row contains
the indexes of the pasients in the group.}
}
\value{
makes a plot
}
\description{
Makes a boxplot for the emperical frequency of the outcome.
}
