% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breslowEstimator.R
\name{breslowEstimator}
\alias{breslowEstimator}
\title{Breslow estimator}
\usage{
breslowEstimator(plpData, timeInDays, parameters)
}
\arguments{
\item{plpData}{a data set like the type used with plp}

\item{timeInDays}{time in days}

\item{parameters}{a data set like the type used with plp}
}
\value{
returns the Breslow estimator
}
\description{
Breslow estimator
}
