% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCoxModel.R
\name{defineCoxModel}
\alias{defineCoxModel}
\title{makes an not fitted model}
\usage{
defineCoxModel(
  coefficients,
  baselinehazard,
  timesofbaselinhazard,
  featureEngineering = NULL
)
}
\arguments{
\item{coefficients}{data set with parameters with rows called betas and covariateIds}

\item{baselinehazard}{number of persons in the returned data set}

\item{timesofbaselinhazard}{data set with parameters with rows called betas and covariateIds}

\item{featureEngineering}{the featureEngineering}
}
\value{
returns a unfitted model
}
\description{
makes an not fitted model
}
