% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empOutcome.R
\name{empOutcome}
\alias{empOutcome}
\title{Title}
\usage{
empOutcome(data, parameter, NoOfCov, Indexes)
}
\arguments{
\item{data}{original dataset}

\item{parameter}{parameters used for genereting new outcomes}

\item{NoOfCov}{number/index of the covariate to study}

\item{Indexes}{matrix with the amount of colums for the simulations, and on the rows the indexes of pasients}
}
\value{
make a plot
}
\description{
Title
}
