% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expBeta_lp.R
\name{expBeta_lp}
\alias{expBeta_lp}
\title{hello}
\usage{
expBeta_lp(TrainingSet, plpData, populationSettings, plpModel)
}
\arguments{
\item{TrainingSet}{a data set like the type used with plp}

\item{plpData}{the plpData}

\item{populationSettings}{the populationSettings}

\item{plpModel}{a data set like the type used with plp}
}
\value{
returns the expBetaz orderd by rowId
}
\description{
hello
}
