% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expBetaZ.R
\name{expBetaZ}
\alias{expBetaZ}
\title{Breslow estimator}
\usage{
expBetaZ(plpData, plpModel)
}
\arguments{
\item{plpData}{a data set like the type used with plp}

\item{plpModel}{a data set like the type used with plp}
}
\value{
returns the expBetaz orderd by rowId
}
\description{
Breslow estimator
}
