% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expLp.R
\name{expLp}
\alias{expLp}
\title{hello}
\usage{
expLp(plpData, plpModel)
}
\arguments{
\item{plpData}{a data set like the type used with plp}

\item{plpModel}{a data set like the type used with plp}
}
\value{
returns the expBetaz orderd by rowId
}
\description{
hello
}
