% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCensoring.R
\name{fitCensoring}
\alias{fitCensoring}
\title{predict censoring times}
\usage{
fitCensoring(Trainingset, modelSettings)
}
\arguments{
\item{propMatrix}{matrix with the probabailies for outcomes}

\item{propMatrixCensoring}{matrix with the probabailies for censoring}

\item{number}{number of people to draw.}

\item{uniqueTimes}{vector with the possible outcome times.}
}
\value{
returns a data set with new outcomes
}
\description{
predict censoring times
}
