% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqInGroup.R
\name{freqInGroup}
\alias{freqInGroup}
\title{frequency for outcome in a specific group}
\usage{
freqInGroup(data, parameters, NoOfCov)
}
\arguments{
\item{data}{orginal dataset}

\item{parameters}{parameters only used to obtain the names of the covariates. this should be different}

\item{NoOfCov}{numer/index of the covariate}
}
\value{
returns a dataframe with the frequencies of the outcome in the different groups
}
\description{
frequency for outcome in a specific group
}
