% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyCovariatePlot.R
\name{frequencyCovariatePlot}
\alias{frequencyCovariatePlot}
\title{box plot of frequency in the outcomes}
\usage{
frequencyCovariatePlot(
  tempdata,
  noSyms,
  noPatientsInSym,
  covariateToStudy,
  plpData,
  noplot,
  colour = "grey"
)
}
\arguments{
\item{covariateToStudy}{a covaraite id}

\item{plpData}{the covariate Data}

\item{colour}{something that ggplot accepts as a colour.}

\item{Data}{dataframe}
}
\value{
returns a boxplot of the frequencies of the covariates in the dataset.
}
\description{
box plot of frequency in the outcomes
}
