% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencyOutcomePlot.R
\name{frequencyOutcomePlot}
\alias{frequencyOutcomePlot}
\title{box plot of frequency in the outcomes}
\usage{
frequencyOutcomePlot(
  OutcomeData,
  noSyms,
  noPatientsInSym,
  plpData,
  placeInPlot,
  colour = "grey"
)
}
\arguments{
\item{OutcomeData}{the data you want visualised.}

\item{noSyms}{number of frequencies to be calculated}

\item{noPatientsInSym}{number of patients to calculate the frequency with.}

\item{plpData}{plpData used for finding the green dots}

\item{placeInPlot}{the place in the plot you want this part to be}

\item{colour}{something that ggplot accepts as a colour.}
}
\value{
returns a boxplot of the frequencies of of outcome for the whole dataset.
}
\description{
box plot of frequency in the outcomes
}
