% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenLine.R
\name{greenLine}
\alias{greenLine}
\title{visual for a covariate}
\usage{
greenLine(plpData, studyCovariateId, youroutcomeId)
}
\arguments{
\item{plpData}{a data set like the type used in PatientLevelPredict}

\item{studyCovariateId}{a number, the covariateId of covariate to restrict to.}

\item{youroutcomeId}{the outcomeID u specified.}
}
\value{
a numer that is used for visualising.
}
\description{
visual for a covariate
}
