% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KaplanMeierPlot.R
\name{KaplanMeierPlot}
\alias{KaplanMeierPlot}
\title{predict censoring times}
\usage{
KaplanMeierPlot(Data, colour = "grey")
}
\arguments{
\item{Data}{dataframe}
}
\value{
returns KaplanMeier plot
}
\description{
predict censoring times
}
