% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KaplanMeierPlotCovariate.R
\name{KaplanMeierPlotCovariate}
\alias{KaplanMeierPlotCovariate}
\title{predict censoring times}
\usage{
KaplanMeierPlotCovariate(
  Data,
  covariateToStudy,
  covariateData,
  colour = "grey"
)
}
\arguments{
\item{Data}{dataframe}

\item{covariateToStudy}{a covariate id the patients are restricted the patients that have this covariate}

\item{colour}{something that ggplot accepts as a colour.}
}
\value{
returns KaplanMeier plot works with ggplot
}
\description{
predict censoring times
}
