% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNewPropsParameters.R
\name{logNewPropsParameters}
\alias{logNewPropsParameters}
\title{simulate a new outcome}
\usage{
logNewPropsParameters(plpData, parameters)
}
\arguments{
\item{plpData}{a plp data set}

\item{parameters}{specifies the parameters used to model new data}
}
\value{
returns a dataframe with propbabities
}
\description{
simulate a new outcome
}
