% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNewPropsParametersPlpModel.R
\name{logNewPropsParametersPlpModel}
\alias{logNewPropsParametersPlpModel}
\title{simulate a new outcome}
\usage{
logNewPropsParametersPlpModel(plpModel, plpData)
}
\arguments{
\item{plpData}{a data set like the type used for plp}

\item{plpResult}{a data set returned by plp}

\item{parameters}{a data set like the type returned by plp}
}
\value{
returns a dataframe with probabilites
}
\description{
simulate a new outcome
}
