% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNewPropsParametersPlpResult.R
\name{logNewPropsParametersPlpResult}
\alias{logNewPropsParametersPlpResult}
\title{simulate a new outcome}
\usage{
logNewPropsParametersPlpResult(
  plpResult,
  plpData,
  parameters = plpResult$covariateSummary$covariateValue
)
}
\arguments{
\item{plpResult}{a runPlp}

\item{parameters}{specifies the parameters used to model new data}

\item{covariates}{should be removed}
}
\value{
returns a dataframe with newOutcomes and subjectId
}
\description{
simulate a new outcome
}
