% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCoxModel.R
\name{makeCoxModel}
\alias{makeCoxModel}
\title{makes an not fitted model}
\usage{
makeCoxModel(coefficients, baselinehazard, timesofbaselinhazard)
}
\arguments{
\item{parameters}{data set with parameters with rows called betas and covariateIds}

\item{modelname}{number of persons in the returned data set}
}
\value{
returns a dataframe with newOutcomes and subjectId
}
\description{
makes an not fitted model
}
