% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLogisticModel.R
\name{makeLogisticModel}
\alias{makeLogisticModel}
\title{makes an not fitted model}
\usage{
makeLogisticModel(parameters)
}
\arguments{
\item{parameters}{data set with parameters with rows called betas and covariateIds}
}
\value{
returns a 'fake' model
}
\description{
makes an not fitted model
}
