% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModel.R
\name{makeModel}
\alias{makeModel}
\title{makes an not fitted model}
\usage{
makeModel(parameters, modelname = "logistic")
}
\arguments{
\item{props}{a vector with the probabilities order by rowId}

\item{noPersons}{number of persons in the returned data set}
}
\value{
returns a dataframe with newOutcomes and subjectId
}
\description{
makes an not fitted model
}
