% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTrainingSet.R
\name{MakeTraingSet}
\alias{MakeTraingSet}
\title{hello}
\usage{
MakeTraingSet(
  plpData,
  executeSettings,
  populationSettings,
  splitSettings,
  sampleSettings,
  preprocessSettings,
  featureEngineeringSettings,
  outcomeId
)
}
\arguments{
\item{plpData}{the plpData}

\item{executeSettings}{the executeSettings}

\item{populationSettings}{the populationSettings}

\item{splitSettings}{the splitSettings}

\item{sampleSettings, }{the sample Settings}

\item{preprocessSettings, }{the preprocessSettings}

\item{featureEngineeringSettings}{the featureEngineeringSettings.}

\item{outcomeId}{the outcomeID}
}
\value{
returns the prediction with an extra column
}
\description{
hello
}
