% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matBoxplot.R
\name{matBoxplot}
\alias{matBoxplot}
\title{Generates a random matrix.
This is usefull for plotting.}
\usage{
matBoxplot(Nofsim = 10, Nofgroups = 200, Nobserv = 30510)
}
\arguments{
\item{Nofsim}{number of rows}

\item{Nofgroups}{number of colums}

\item{Nobserv}{the maximum number in the matrix}
}
\value{
a random matrix
}
\description{
Generates a random matrix.
This is usefull for plotting.
}
