% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matBoxplot2.R
\name{matBoxplot2}
\alias{matBoxplot2}
\title{usefull for visualizing}
\usage{
matBoxplot2(Nofsim = 10, Nofgroups = 200, Nobserv = 30510)
}
\arguments{
\item{Nofsim}{number of simulations/ number of columns}

\item{Nofgroups}{number of groups/ number of rows}

\item{Nobserv}{maximum value in the matrix}
}
\value{
a random matrix with identical columns
}
\description{
usefull for visualizing
}
