% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newData.R
\name{newData}
\alias{newData}
\title{Title}
\usage{
newData(data, parameter, Indexes)
}
\arguments{
\item{data}{orginal dataset}

\item{parameter}{parameters used to obtain new data}

\item{Indexes}{a vector of indexes of the patients u want new outcomes for}
}
\value{
returns a dataframe with newOutcomes and indexes
}
\description{
Title
}
