% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newGroupData.R
\name{newGroupData}
\alias{newGroupData}
\title{Generates a new group with new outcomes
that follows the specified model.}
\usage{
newGroupData(data, N, parameter)
}
\arguments{
\item{data}{Original dataset}

\item{N}{number of elements in your new group}

\item{parameter}{parameter for new generated outcomes}
}
\value{
returns a dataframe(with N rows) with ind for indexes, and newOutcomes
}
\description{
Generates a new group with new outcomes
that follows the specified model.
}
