% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newOutcomes.R
\name{newOutcomes}
\alias{newOutcomes}
\title{simulate a new outcome}
\usage{
newOutcomes(noPersons, props)
}
\arguments{
\item{noPersons}{number of persons in the returned data set}

\item{props}{a vector with the probabilities ordered by rowId}
}
\value{
returns a dataframe with newOutcomes and subjectId
}
\description{
simulate a new outcome
}
