% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newOutcomesCensoredSurvivalTimes.R
\name{newOutcomesCensoredSurvivalTimes}
\alias{newOutcomesCensoredSurvivalTimes}
\title{simulating of a survival time}
\usage{
newOutcomesCensoredSurvivalTimes(
  plpModelCensoring,
  expbetasCensor,
  plpModel,
  expbetas,
  number
)
}
\arguments{
\item{plpModelCensoring}{a plpModel for the censoringtimes}

\item{expbetasCensor}{a vector with the expbeta for the censoring model}

\item{plpModel}{a plpModel for the outcomes}

\item{expbetas}{the expbetas for the outcomemodel}

\item{number}{number of people to draw.}
}
\value{
returns a data set with new outcomes
}
\description{
simulating of a survival time
}
