% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newOutcomesParameters.R
\name{newOutcomesParameters}
\alias{newOutcomesParameters}
\title{simulate a new outcome}
\usage{
newOutcomesParameters(
  plpResult,
  covariates,
  noSimulations,
  noPersons,
  parameters = plpResult$covariateSummary$covariateValue
)
}
\arguments{
\item{plpResult}{a runPlp}

\item{noSimulations}{number of dataset one wants}

\item{noPersons}{number of persons in each data set}

\item{parameters}{specifies the parameters used to model new data}
}
\value{
returns a dataframe with newOutcomes and subjectId
}
\description{
simulate a new outcome
}
