% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newOutcomesSurvivalTimes.R
\name{newOutcomesSurvivalTimes}
\alias{newOutcomesSurvivalTimes}
\title{simulating of a survival time}
\usage{
newOutcomesSurvivalTimes(propMatrix, number, uniqueTimes)
}
\arguments{
\item{propMatrix}{matrix with the probabailies}

\item{number}{number of people to draw.}

\item{uniqueTimes}{vector with the possible outcome times.}
}
\value{
returns a data set with new outcomes
}
\description{
simulating of a survival time
}
