% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPropsCox.R
\name{newPropsCox}
\alias{newPropsCox}
\title{New probablies for cox model}
\usage{
newPropsCox(plpModel, plpData, timeInDays)
}
\arguments{
\item{plpModel}{a data set like the type used with plp}

\item{plpData}{a data set like the type used with plp}

\item{timeInDays}{time in days}
}
\value{
returns the Breslow estimator
}
\description{
New probablies for cox model
}
