% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPropsCoxversion2.R
\name{newPropsCoxversion2}
\alias{newPropsCoxversion2}
\title{New probablies for cox model}
\usage{
newPropsCoxversion2(plpModel, plpData)
}
\arguments{
\item{plpModel}{a data set like the type used with plp}

\item{plpData}{a data set like the type used with plp}
}
\value{
returns the Breslow estimator
}
\description{
New probablies for cox model
}
