% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPropsParameters.R
\name{newPropsParameters}
\alias{newPropsParameters}
\title{simulate a new outcome}
\usage{
newPropsParameters(plpData, parameters, modelName)
}
\arguments{
\item{plpData}{a data set as used for plp}

\item{parameters}{parameters as returned by plp}

\item{modelName}{now it only works for "logistic"}
}
\value{
returns a dataframe with probabilties d
}
\description{
simulate a new outcome
}
