% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPropsParametersPlpResults.R
\name{newPropsParametersPlpResults}
\alias{newPropsParametersPlpResults}
\title{simulate a new outcome}
\usage{
newPropsParametersPlpResults(
  plpResult,
  plpData,
  parameters = plpResult$covariateSummary$covariateValue
)
}
\arguments{
\item{plpResult}{a runPlp}

\item{parameters}{specifies the parameters used to model new data}

\item{noSimulations}{number of dataset one wants}

\item{noPersons}{number of persons in each data set}
}
\value{
returns a dataframe with newOutcomes and subjectId
}
\description{
simulate a new outcome
}
