% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observedFrequency.R
\name{observedFrequency}
\alias{observedFrequency}
\title{simulate a new outcome}
\usage{
observedFrequency(data, plpData, studyCovariateId)
}
\arguments{
\item{data}{A data set as returned by newOutcomes}

\item{plpData}{a dataset}

\item{studyCovariateId}{the covariateId of the to study covariate}
}
\value{
returns a vector with length of number of simulations with the frequency
of the outcome for the pasients with covariate of the specified covariateId
}
\description{
simulate a new outcome
}
