% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiffInTotalOutcomes.R
\name{plotDiffInTotalOutcomes}
\alias{plotDiffInTotalOutcomes}
\title{plots the difference bewteen the total of outcomes in the original data set
and a data set simulated with the same length}
\usage{
plotDiffInTotalOutcomes(data, N, parameter)
}
\arguments{
\item{data}{orginal dataset}

\item{N}{number of simulations}

\item{parameter}{parameters used for generating the data}
}
\value{
returns a plot
}
\description{
plots the difference bewteen the total of outcomes in the original data set
and a data set simulated with the same length
}
