% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probsUnderModel.R
\name{probsUnderModel}
\alias{probsUnderModel}
\title{returns the probability of an outcome under the parameters specified}
\usage{
probsUnderModel(data, parameters, index = 1:dim(data[1])[1])
}
\arguments{
\item{data}{orginal dataset}

\item{parameters}{parameters}

\item{index}{optional: indexes for the elements u want the probability of}
}
\value{
returns a vector with probabilities
}
\description{
returns the probability of an outcome under the parameters specified
}
