% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlasmode.R
\name{runPlasmode}
\alias{runPlasmode}
\title{simuates a new dataset.}
\usage{
runPlasmode(
  plpdata,
  outcomeId,
  populationSettings,
  splitSettings,
  sampleSettings,
  featureEngineeringSettings,
  preprocessSettings,
  modelSettings,
  executeSettings,
  numberToSimulate
)
}
\arguments{
\item{plpdata}{the plpdata}

\item{outcomeId}{the outcome id}

\item{populationSettings}{as made with createStudyPopulationSettings()}

\item{splitSettings}{as made with createDefaultSplitSetting()}

\item{sampleSettings}{as made with createSampleSettings ()}

\item{featureEngineeringSettings}{as made with createFeatureEngineeringSettings()}

\item{preprocessSettings}{as made with createPreprocessSettings()}

\item{modelSettings}{specifies your chosen model like setCoxModel()}

\item{executeSettings}{as made with createExecuteSettings()}

\item{numberToSimulate}{number you choose # should in the future be sample settings}
}
\value{
returns a new dataset
}
\description{
simuates a new dataset.
}
