% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSurvivaltimes.R
\name{simulateSurvivaltimes}
\alias{simulateSurvivaltimes}
\title{simulating of a survival time}
\usage{
simulateSurvivaltimes(plpModel, plpData, number, populationSettings)
}
\arguments{
\item{plpModel}{model with}

\item{number}{number of people to draw.}

\item{expbetas}{expbetas.}
}
\value{
returns a data set with new outcomes
}
\description{
simulating of a survival time
}
