% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSurvivaltimesCohortMethod.R
\name{simulateSurvivaltimesCohortMethod}
\alias{simulateSurvivaltimesCohortMethod}
\title{simulating of a survival time}
\usage{
simulateSurvivaltimesCohortMethod(
  outcomeModel,
  cmData,
  numberToSimulate,
  population,
  populationSettings
)
}
\arguments{
\item{numberToSimulate}{number of people to draw and simulate new outcomes for.}

\item{population}{the population to draw from}

\item{populationSettings}{the populationSettings}

\item{plpModel}{a plpModel}

\item{plpData}{a plpdata}
}
\value{
returns a data set with new outcomes
}
\description{
simulating of a survival time
}
