% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSurvivaltimesWithCensoring.R
\name{simulateSurvivaltimesWithCensoring}
\alias{simulateSurvivaltimesWithCensoring}
\title{predict censoring times}
\usage{
simulateSurvivaltimesWithCensoring(
  censorModel,
  plpData,
  population,
  populationSettings,
  numberToSimulate
)
}
\arguments{
\item{censorModel}{a Model as made by fitModelWithCensoring a list with twomodels named censorModel and outcomesModel}

\item{plpData}{the plpData}

\item{population}{the population to drawfrom}

\item{populationSettings}{the populationSettings}

\item{numberToSimulate}{number of people to simulate new outcomes for.}
}
\value{
returns a data set with new outcomes or censoring times
}
\description{
predict censoring times
}
