% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSurvivaltimesWithCensoringCovariate.R
\name{simulateSurvivaltimesWithCensoringCovariate}
\alias{simulateSurvivaltimesWithCensoringCovariate}
\title{predict censoring times}
\usage{
simulateSurvivaltimesWithCensoringCovariate(
  censorModel,
  plpData,
  population,
  populationSettings,
  numberToSimulate,
  covariateToStudy
)
}
\arguments{
\item{censorModel}{a Model as made by fitModelWithCensoring a list with twomodels named censorModel and outcomesModel}

\item{plpData}{the plpData}

\item{population}{the population to drawfrom}

\item{populationSettings}{the populationSettings}

\item{numberToSimulate}{number of people to simulate new outcomes for.}

\item{covariateToStudy}{the covariate Id that has to be present in the simulated data}
}
\value{
returns a data set with new outcomes or censoring times
}
\description{
predict censoring times
}
