% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateWithCensoring.R
\name{simulateWithCensoring}
\alias{simulateWithCensoring}
\title{predict censoring times}
\usage{
simulateWithCensoring(fitCensor, plpData, populationSettings, number)
}
\arguments{
\item{number}{number of people to draw.}

\item{propMatrix}{matrix with the probabailies for outcomes}

\item{propMatrixCensoring}{matrix with the probabailies for censoring}

\item{uniqueTimes}{vector with the possible outcome times.}
}
\value{
returns a data set with new outcomes
}
\description{
predict censoring times
}
