% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalPlotCox.R
\name{SurvivalPlotCox}
\alias{SurvivalPlotCox}
\title{predict censoring times}
\usage{
SurvivalPlotCox(plpModel, colour = "red")
}
\arguments{
\item{plpModel}{the plpmodel that has an baselinesurvival}

\item{colour}{one can specify the colour just as in ggplot.}
}
\value{
adds the baselinesurvival from the model to the plot
}
\description{
predict censoring times
}
