% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoreticalExpectation.R
\name{theoreticalExpectation}
\alias{theoreticalExpectation}
\title{frequency of a covariate in a dataset}
\usage{
theoreticalExpectation(plpData, parameters)
}
\arguments{
\item{plpData}{a data set}

\item{parameters}{parameters in a dataset, dataset has two rows first with the parameter and the second with the covariateId}
}
\value{
returns a number between 0 and 1 what is the estimated frequency,
}
\description{
frequency of a covariate in a dataset
}
