% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual1.R
\name{Visual1}
\alias{Visual1}
\title{simulate a new outcome}
\usage{
Visual1(
  plpResult,
  covariates,
  studyCovariateId,
  noSimulations,
  noPersons,
  parameters = plpResult$covariateSummary$covariateValue
)
}
\arguments{
\item{data}{a data set as returned by newOutcomesParameters}
}
\value{
returns a boxplot for the frequencies of the outcome
}
\description{
simulate a new outcome
}
