% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visual2.R
\name{Visual2}
\alias{Visual2}
\title{simulate a new outcome}
\usage{
Visual2(data, covariates, studyCovariateId)
}
\arguments{
\item{data}{a data set as returned by newOutcomesParameters}
}
\value{
returns a plot
}
\description{
simulate a new outcome
}
