% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualOutcome.R
\name{visualOutcome}
\alias{visualOutcome}
\title{simulate a new outcome}
\usage{
visualOutcome(plpData, noSimulations, noPersons, parameters)
}
\arguments{
\item{plpData}{a data set like the type used for plp}

\item{noSimulations}{number of simulations}

\item{noPersons}{number of persons}

\item{parameters}{a data set like the type used for plp}
}
\value{
returns a histogram for the frequencies of the outcome
}
\description{
simulate a new outcome
}
