% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualOutcomeCovariateId.R
\name{visualOutcomeCovariateId}
\alias{visualOutcomeCovariateId}
\title{simulate a new outcome}
\usage{
visualOutcomeCovariateId(
  plpResult,
  plpData,
  studyCovariateId,
  noSimulations,
  noPersons,
  parameters = plpResult$covariateSummary$covariateValue
)
}
\arguments{
\item{plpResult}{a data set returned by plp}

\item{plpData}{a data set like the type used for plp}

\item{studyCovariateId}{a number, the covariateId of the to study covariate}

\item{noSimulations}{number of simulations}

\item{noPersons}{number of persons}

\item{parameters}{a data set like the type used for plp}
}
\value{
returns a histogram for the frequencies of the outcome
}
\description{
simulate a new outcome
}
