% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualOutcomeCovariateId2.R
\name{visualOutcomeCovariateId2}
\alias{visualOutcomeCovariateId2}
\title{simulate a new outcome}
\usage{
visualOutcomeCovariateId2(
  plpData,
  restrictToCovariateId,
  noSimulations,
  noPersons,
  parameters,
  modelName
)
}
\arguments{
\item{plpData}{a data set like the type used for plp}

\item{noSimulations}{number of simulations}

\item{noPersons}{number of persons}

\item{parameters}{a data set like the type used for plp}

\item{plpResult}{a data set returned by plp}

\item{studyCovariateId}{a number, the covariateId of the to study covariate}
}
\value{
returns a histogram for the frequencies of the outcome
}
\description{
simulate a new outcome
}
