% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustBaselineSurvival.R
\name{AdjustBaselineSurvival}
\alias{AdjustBaselineSurvival}
\title{adjust an Baseline Survival function such that the overal event rate is fixed}
\usage{
AdjustBaselineSurvival(
  BaselineSurv,
  timeTofixat,
  proptofixwith,
  expbetaz,
  intervalSolution = c(-100, 100)
)
}
\arguments{
\item{BaselineSurv}{a data frame with the the colums called surv and time.}

\item{timeTofixat}{a time in days to fix the eventrate at. has to be in the times of the baseline survival}

\item{proptofixwith}{probabilitie of the event something somethign}

\item{expbetaz}{the vector with expbetasz made with function expBetaZ()}

\item{intervalSolution}{the inteval used for finding the solution.}
}
\value{
returns a modified Baselinesurvivalfunction
}
\description{
adjust an Baseline Survival function such that the overal event rate is fixed
}
