% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlotfreq.R
\name{boxplotFreq}
\alias{boxplotFreq}
\title{Makes a boxplot for the emperical frequency of the outcome.}
\usage{
boxplotFreq(data, parameter, noOfCov, indexes)
}
\arguments{
\item{data}{Original dataset}

\item{parameter}{Parameters for the model used to generate the datasets}

\item{NoOfCov}{Number/place of the covariate u want study with the plot}

\item{Indexes}{Matrix with for echt colum a simulation and each row contains
the indexes of the pasients in the group.}
}
\value{
makes a plot
}
\description{
Makes a boxplot for the emperical frequency of the outcome.
}
