% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequencyOutcomeCovariatePlot.R
\name{frequencyOutcomeCovariatePlot}
\alias{frequencyOutcomeCovariatePlot}
\title{box plot of frequency in a covariate the outcomes}
\usage{
frequencyOutcomeCovariatePlot(
  tempdata,
  noPatientsInSym,
  noSyms,
  covariateToStudy,
  plpData,
  noplot,
  colour = "grey"
)
}
\arguments{
\item{covariateToStudy}{a covaraite id}

\item{plpData}{the covariate Data}

\item{colour}{something that ggplot accepts as a colour.}

\item{Data}{dataframe}
}
\value{
returns a boxplot of the frequencie of outcome when a specific covariates is present.
}
\description{
box plot of frequency in a covariate the outcomes
}
