% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FrequencyOutcomePlot.R
\name{frequencyOutcomePlot}
\alias{frequencyOutcomePlot}
\title{box plot of frequency in the outcomes}
\usage{
frequencyOutcomePlot(
  tempdata,
  noSyms,
  noPatientsInSym,
  noplot,
  colour = "grey"
)
}
\arguments{
\item{colour}{something that ggplot accepts as a colour.}

\item{Data}{dataframe}
}
\value{
returns a boxplot of the frequencies of of outcome for the whole dataset.
}
\description{
box plot of frequency in the outcomes
}
