% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KaplanMeierPlotFilterCovariate.R
\name{KaplanMeierPlotFilterCovariate}
\alias{KaplanMeierPlotFilterCovariate}
\title{predict censoring times}
\usage{
KaplanMeierPlotFilterCovariate(
  Data,
  covariateToStudy,
  plpData,
  colour = "grey"
)
}
\arguments{
\item{Data}{dataframe}

\item{covariateToStudy}{a covaraite id}

\item{plpData}{the covariate Data}

\item{colour}{something that ggplot accepts as a colour.}
}
\value{
returns KaplanMeier plot works with ggplot the data has been filtered
such that only the covariateToStudy is present
}
\description{
predict censoring times
}
