% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logNewPropsParameters.R
\name{logNewPropsParameters}
\alias{logNewPropsParameters}
\title{simulate a new outcome}
\usage{
logNewPropsParameters(plpData, parameters)
}
\arguments{
\item{parameters}{specifies the parameters used to model new data}

\item{plpResult}{a runPlp}

\item{covariates}{should be removed}
}
\value{
returns a dataframe with newOutcomes and subjectId
}
\description{
simulate a new outcome
}
