% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTrainingSet.R
\name{MakeTraingSet}
\alias{MakeTraingSet}
\title{hello}
\usage{
MakeTraingSet(
  plpData,
  executeSettings,
  populationSettings,
  splitSettings,
  sampleSettings,
  preprocessSettings,
  featureEngineeringSettings,
  outcomeId
)
}
\arguments{
\item{prediction}{form a predictPlp}

\item{baselineSurvival}{}
}
\value{
returns the prediction with an extra column
}
\description{
hello
}
