% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newOutcomes.R
\name{newOutcomes}
\alias{newOutcomes}
\title{simulate a new outcome}
\usage{
newOutcomes(noPersons, props)
}
\arguments{
\item{noPersons}{number of persons in each data set}

\item{plpResult}{a runPlp}

\item{noSimulations}{number of dataset one wants}
}
\value{
returns a dataframe with newOutcomes and subjectId
}
\description{
simulate a new outcome
}
