% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newOutcomesCensoredSurvivalTimes.R
\name{newOutcomesCensoredSurvivalTimes}
\alias{newOutcomesCensoredSurvivalTimes}
\title{simulating of a survival time}
\usage{
newOutcomesCensoredSurvivalTimes(
  plpModelCensoring,
  expbetasCensor,
  plpModel,
  expbetas,
  number
)
}
\arguments{
\item{number}{number of people to draw.}

\item{propMatrix}{matrix with the probabailies for outcomes}

\item{propMatrixCensoring}{matrix with the probabailies for censoring}

\item{uniqueTimes}{vector with the possible outcome times.}
}
\value{
returns a data set with new outcomes
}
\description{
simulating of a survival time
}
