% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newOutcomesSurvivalTimes.R
\name{newOutcomesSurvivalTimes}
\alias{newOutcomesSurvivalTimes}
\title{simulating of a survival time}
\usage{
newOutcomesSurvivalTimes(plpModel, expbetas, number)
}
\arguments{
\item{plpModel}{model with}

\item{expbetas}{expbetas.}

\item{number}{number of people to draw.}
}
\value{
returns a data set with new outcomes
}
\description{
simulating of a survival time
}
