% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPropsParameters.R
\name{newPropsParameters}
\alias{newPropsParameters}
\title{simulate a new outcome}
\usage{
newPropsParameters(plpData, parameters, modelName)
}
\arguments{
\item{parameters}{specifies the parameters used to model new data}

\item{plpResult}{a runPlp}

\item{noSimulations}{number of dataset one wants}

\item{noPersons}{number of persons in each data set}
}
\value{
returns a dataframe with newOutcomes and subjectId
}
\description{
simulate a new outcome
}
