% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPropsParametersPlpModel.R
\name{newPropsParametersPlpModel}
\alias{newPropsParametersPlpModel}
\title{simulate a new outcome}
\usage{
newPropsParametersPlpModel(plpModel, plpData)
}
\arguments{
\item{plpData}{a data set like the type used for plp}

\item{plpResult}{a data set returned by plp}

\item{parameters}{a data set like the type used for plp}
}
\value{
returns a dataframe with probabilites
}
\description{
simulate a new outcome
}
