% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPropsParametersPlpResults.R
\name{newPropsParametersPlpResults}
\alias{newPropsParametersPlpResults}
\title{simulate a new outcome}
\usage{
newPropsParametersPlpResults(
  plpResult,
  plpData,
  parameters = plpResult$covariateSummary$covariateValue
)
}
\arguments{
\item{plpResult}{a data set returned by plp}

\item{plpData}{a data set like the type used for plp}

\item{parameters}{a data set like the type used for plp}
}
\value{
returns a dataframe with probabilites
}
\description{
simulate a new outcome
}
