% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observedFrequencies.R
\name{observedFrequencies}
\alias{observedFrequencies}
\title{simulate a new outcome}
\usage{
observedFrequencies(data, covariates, studyCovariateId)
}
\arguments{
\item{data}{A data set as returned by newOutcomesParameters}

\item{covariates}{the covariates (class covariateData)}

\item{studycovariateId}{the covariateId of the to study covariate}
}
\value{
returns a vector with length of number of simulations with the frequency
of the outcome for the pasients with covariate of the specified covariateId
}
\description{
simulate a new outcome
}
