% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observedFrequency.R
\name{observedFrequency}
\alias{observedFrequency}
\title{simulate a new outcome}
\usage{
observedFrequency(data, plpData, studyCovariateId)
}
\arguments{
\item{data}{A data set as returned by newOutcomes}

\item{plpData}{a data set like the type used for plp}

\item{studyCovariateId}{the covariateId of the covariate for wich the frequency is wanted}
}
\value{
the frequency of the outcome for the patients with the specified covariateId present
}
\description{
simulate a new outcome
}
