% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSurvivaltimes.R
\name{simulateSurvivaltimes}
\alias{simulateSurvivaltimes}
\title{simulating of a survival time}
\usage{
simulateSurvivaltimes(
  plpModel,
  plpData,
  numberToSimulate,
  population,
  populationSettings
)
}
\arguments{
\item{plpModel}{a plpModel}

\item{plpData}{a plpdata}

\item{numberToSimulate}{number of people to draw and simulate new outcomes for.}

\item{population}{the population to draw from}

\item{populationSettings}{the populationSettings}
}
\value{
returns a data set with new outcomes
}
\description{
simulating of a survival time
}
