% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSurvivaltimesWithCensoring.R
\name{simulateSurvivaltimesWithCensoring}
\alias{simulateSurvivaltimesWithCensoring}
\title{predict censoring times}
\usage{
simulateSurvivaltimesWithCensoring(
  fitCensor,
  plpData,
  populationSettings,
  number
)
}
\arguments{
\item{number}{number of people to w.}

\item{propMatrix}{matrix with the probabailies for outcomes}

\item{propMatrixCensoring}{matrix with the probabailies for censoring}

\item{uniqueTimes}{vector with the possible outcome times.}
}
\value{
returns a data set with new outcomes
}
\description{
predict censoring times
}
