% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalPlotCox.R
\name{SurvivalPlotCox}
\alias{SurvivalPlotCox}
\title{predict censoring times}
\usage{
SurvivalPlotCox(plpModel, colour = "red")
}
\value{
returns survivalfunction ro plot
}
\description{
predict censoring times
}
