% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoreticalExpectation.R
\name{theoreticalExpectation}
\alias{theoreticalExpectation}
\title{frequency of a covariate in a dataset}
\usage{
theoreticalExpectation(plpData, parameters)
}
\arguments{
\item{plpData}{a data set}

\item{covariateId}{number of a covariate id}
}
\value{
returns a number between 0 and 1
}
\description{
frequency of a covariate in a dataset
}
