% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualOutcome.R
\name{visualOutcome}
\alias{visualOutcome}
\title{simulate a new outcome}
\usage{
visualOutcome(
  plpResult,
  plpData,
  noSimulations,
  noPersons,
  parameters = plpResult$covariateSummary$covariateValue
)
}
\arguments{
\item{data}{a data set as returned by newOutcomesParameters}
}
\value{
returns a boxplot for the frequencies of the outcome
}
\description{
simulate a new outcome
}
