% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualOutcomeCovariateId.R
\name{visualOutcomeCovariateId}
\alias{visualOutcomeCovariateId}
\title{simulate a new outcome}
\usage{
visualOutcomeCovariateId(
  plpResult,
  plpData,
  studyCovariateId,
  noSimulations,
  noPersons,
  parameters = plpResult$covariateSummary$covariateValue
)
}
\arguments{
\item{data}{a data set as returned by newOutcomesParameters}
}
\value{
returns a boxplot for the frequencies of the outcome
}
\description{
simulate a new outcome
}
