% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualOutcomeCovariateId.R
\name{visualOutcomeCovariateId}
\alias{visualOutcomeCovariateId}
\title{visual for a covariate}
\usage{
visualOutcomeCovariateId(
  plpData,
  studyCovariateId,
  noSimulations,
  noPersons,
  parameters
)
}
\arguments{
\item{plpData}{a data set like the type used in PatientLevelPredict}

\item{studyCovariateId}{a number, the covariateId of covariate to restrict to.}

\item{noSimulations}{number of simulations}

\item{noPersons}{number of persons in each simulation}

\item{parameters}{a data set with parameters like the type used for plp}
}
\value{
returns a histogram for the frequencies of the outcome the frequency is calculated with only the patients that have the given covariate Id present.
}
\description{
visual for a covariate
}
