% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualOutcomeCovariateId2.R
\name{visualOutcomeCovariateId2}
\alias{visualOutcomeCovariateId2}
\title{visual for a covariate}
\usage{
visualOutcomeCovariateId2(
  plpData,
  restrictToCovariateId,
  noSimulations,
  noPersons,
  parameters,
  modelName
)
}
\arguments{
\item{plpData}{a data set like the type used in PatientLevelPredict}

\item{restrictToCovariateId}{a number, the covariateId of covariate to restrict to.}

\item{noSimulations}{number of simulations}

\item{noPersons}{number of persons in each simulation}

\item{parameters}{a data set with parameters like the type used for plp}

\item{modelName}{a string with the type of model.}
}
\value{
returns a histogram for the frequencies of the outcome there is only simulated from patients that have the given covariate Id.
}
\description{
visual for a covariate
}
