% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualOutcomePlpResult.R
\name{visualOutcomePlpResult}
\alias{visualOutcomePlpResult}
\title{simulate a new outcome}
\usage{
visualOutcomePlpResult(
  plpResult,
  plpData,
  noSimulations,
  noPersons,
  parameters = plpResult$covariateSummary$covariateValue
)
}
\arguments{
\item{plpResult}{a data set as returned by newOutcomesParameters}

\item{plpData}{a data set as with covariates,}

\item{noSimulations}{number of simulations preformed}

\item{noPersons}{numner of person in each simulatons}

\item{parameters}{a data set with parameters}
}
\value{
returns a histogram for the frequencies of the outcome
}
\description{
simulate a new outcome
}
